Add ListingPageAccessWrapper component and take it into use
