+++ src/containers/ListingPage/ListingPageAccessWrapper.js
@@ -1,0 +1,90 @@
+import React from 'react';
+import { useHistory, useLocation } from 'react-router-dom';
+
+import { useConfiguration } from '../../context/configurationContext';
+import { useRouteConfiguration } from '../../context/routeConfigurationContext';
+
+import { useIntl } from '../../util/reactIntl';
+import {
+  isErrorNoViewingPermission,
+  isErrorUserPendingApproval,
+  isForbiddenError,
+} from '../../util/errors';
+import {
+  NO_ACCESS_PAGE_USER_PENDING_APPROVAL,
+  NO_ACCESS_PAGE_VIEW_LISTINGS,
+} from '../../util/urlHelpers';
+import { hasPermissionToViewData, isUserAuthorized } from '../../util/userHelpers';
+
+import { NamedRedirect } from '../../components';
+
+/**
+ * Shared access gating for ListingPage variants: forbidden, private marketplace,
+ * pending approval, and no viewing rights. Injects router/context props into the page component.
+ *
+ * @param {Object} props
+ * @param {React.ComponentType} props.PageComponent - Listing page component receiving injected props
+ * @param {Object} props - Remaining props from Redux `connect`
+ * @returns {JSX.Element}
+ */
+const ListingPageAccessWrapper = ({ PageComponent, ...rest }) => {
+  const config = useConfiguration();
+  const routeConfiguration = useRouteConfiguration();
+  const intl = useIntl();
+  const history = useHistory();
+  const location = useLocation();
+
+  const showListingError = rest.showListingError;
+  const isVariant = rest.params?.variant != null;
+  const currentUser = rest.currentUser;
+  if (isForbiddenError(showListingError) && !isVariant && !currentUser) {
+    // This can happen if private marketplace mode is active
+    return (
+      <NamedRedirect
+        name="SignupPage"
+        state={{ from: `${location.pathname}${location.search}${location.hash}` }}
+      />
+    );
+  }
+
+  const isPrivateMarketplace = config.accessControl.marketplace.private === true;
+  const isUnauthorizedUser = currentUser && !isUserAuthorized(currentUser);
+  const hasNoViewingRights = currentUser && !hasPermissionToViewData(currentUser);
+  const hasUserPendingApprovalError = isErrorUserPendingApproval(showListingError);
+
+  if ((isPrivateMarketplace && isUnauthorizedUser) || hasUserPendingApprovalError) {
+    return (
+      <NamedRedirect
+        name="NoAccessPage"
+        params={{ missingAccessRight: NO_ACCESS_PAGE_USER_PENDING_APPROVAL }}
+      />
+    );
+  }
+  if (
+    (hasNoViewingRights && isForbiddenError(showListingError)) ||
+    isErrorNoViewingPermission(showListingError)
+  ) {
+    // If the user has no viewing rights, fetching anything but their own listings
+    // will return a 403 error. If that happens, redirect to NoAccessPage.
+    return (
+      <NamedRedirect
+        name="NoAccessPage"
+        params={{ missingAccessRight: NO_ACCESS_PAGE_VIEW_LISTINGS }}
+      />
+    );
+  }
+
+  return (
+    <PageComponent
+      config={config}
+      routeConfiguration={routeConfiguration}
+      intl={intl}
+      history={history}
+      location={location}
+      showOwnListingsOnly={hasNoViewingRights}
+      {...rest}
+    />
+  );
+};
+
+export default ListingPageAccessWrapper;
